Moved hasIdamToken, hasUserId to CallbackValidator
