+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/CallbackValidations.java
@@ -14,7 +14,6 @@ import java.time.format.DateTimeFormatterBuilder;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import javax.annotation.Nonnull;
 
 import static io.vavr.control.Validation.invalid;
 import static io.vavr.control.Validation.valid;
@@ -41,20 +40,6 @@ public final class CallbackValidations {
         .withZone(ZoneOffset.UTC);
 
     private CallbackValidations() {
-    }
-
-    @Nonnull
-    static Validation<String, String> hasIdamToken(String idamToken) {
-        return idamToken != null
-            ? valid(idamToken)
-            : invalid("Callback has no Idam token received in the header");
-    }
-
-    @Nonnull
-    static Validation<String, String> hasUserId(String userId) {
-        return userId != null
-            ? valid(userId)
-            : invalid("Callback has no user id received in the header");
     }
 
     private static Optional<String> getJourneyClassification(CaseDetails theCase) {
