+++ src/test/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/CallbackValidationsTest.java
@@ -27,8 +27,6 @@ import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.doma
 @SuppressWarnings("checkstyle:LineLength")
 class CallbackValidationsTest {
 
-    public static final String NO_IDAM_TOKEN_RECEIVED_ERROR = "Callback has no Idam token received in the header";
-    public static final String NO_USER_ID_RECEIVED_ERROR = "Callback has no user id received in the header";
     public static final String JOURNEY_CLASSIFICATION = "journeyClassification";
     public static final String CLASSIFICATION_EXCEPTION = "EXCEPTION";
 
@@ -59,56 +57,6 @@ class CallbackValidationsTest {
             expectedValue,
             CallbackValidations::hasJourneyClassificationForAttachToCase,
             error
-        );
-    }
-
-    private static Object[][] idamTokenTestParams() {
-        return new Object[][]{
-            {"null idam token", null, false, NO_IDAM_TOKEN_RECEIVED_ERROR},
-            {"valid idam token", "valid token", true, "valid token"}
-        };
-    }
-
-    @ParameterizedTest(name = "{0}: valid:{2} error/value:{3}")
-    @MethodSource("idamTokenTestParams")
-    @DisplayName("Should have idam token in the request")
-    void idamTokenInTheRequestTest(
-        String caseDescription,
-        String input,
-        boolean valid,
-        String expectedValueOrError
-    ) {
-        checkValidation(
-            input,
-            valid,
-            expectedValueOrError,
-            CallbackValidations::hasIdamToken,
-            expectedValueOrError
-        );
-    }
-
-    private static Object[][] userIdTestParams() {
-        return new Object[][]{
-            {"null user id", null, false, NO_USER_ID_RECEIVED_ERROR},
-            {"valid user id", "valid user id", true, "valid user id"}
-        };
-    }
-
-    @ParameterizedTest(name = "{0}: valid:{2} error/value:{3}")
-    @MethodSource("userIdTestParams")
-    @DisplayName("Should have user id in the request")
-    void userIdInTheRequestTest(
-        String caseDescription,
-        String input,
-        boolean valid,
-        String expectedValueOrError
-    ) {
-        checkValidation(
-            input,
-            valid,
-            expectedValueOrError,
-            CallbackValidations::hasUserId,
-            expectedValueOrError
         );
     }
 
@@ -147,16 +95,6 @@ class CallbackValidationsTest {
                                      boolean valid,
                                      T realValue,
                                      Function<CaseDetails, Validation<String, ?>> validationMethod,
-                                     String errorString) {
-        Validation<String, ?> validation = validationMethod.apply(input);
-
-        softAssertions(valid, realValue, errorString, validation);
-    }
-
-    private <T> void checkValidation(String input,
-                                     boolean valid,
-                                     T realValue,
-                                     Function<String, Validation<String, ?>> validationMethod,
                                      String errorString) {
         Validation<String, ?> validation = validationMethod.apply(input);
