+++ src/test/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/CallbackValidatorTest.java
@@ -40,6 +40,8 @@ class CallbackValidatorTest {
 
     private static final String JOURNEY_CLASSIFICATION = "journeyClassification";
     private static final String CLASSIFICATION_EXCEPTION = "EXCEPTION";
+    private static final String NO_IDAM_TOKEN_RECEIVED_ERROR = "Callback has no Idam token received in the header";
+    private static final String NO_USER_ID_RECEIVED_ERROR = "Callback has no user id received in the header";
 
     @Mock
     private CaseReferenceValidator caseReferenceValidator;
@@ -301,6 +303,56 @@ class CallbackValidatorTest {
         );
     }
 
+    private static Object[][] idamTokenTestParams() {
+        return new Object[][]{
+                {"null idam token", null, false, NO_IDAM_TOKEN_RECEIVED_ERROR},
+                {"valid idam token", "valid token", true, "valid token"}
+        };
+    }
+
+    @ParameterizedTest(name = "{0}: valid:{2} error/value:{3}")
+    @MethodSource("idamTokenTestParams")
+    @DisplayName("Should have idam token in the request")
+    void idamTokenInTheRequestTest(
+            String caseDescription,
+            String input,
+            boolean valid,
+            String expectedValueOrError
+    ) {
+        checkValidation(
+                input,
+                valid,
+                expectedValueOrError,
+                callbackValidator::hasIdamToken,
+                expectedValueOrError
+        );
+    }
+
+    private static Object[][] userIdTestParams() {
+        return new Object[][]{
+                {"null user id", null, false, NO_USER_ID_RECEIVED_ERROR},
+                {"valid user id", "valid user id", true, "valid user id"}
+        };
+    }
+
+    @ParameterizedTest(name = "{0}: valid:{2} error/value:{3}")
+    @MethodSource("userIdTestParams")
+    @DisplayName("Should have user id in the request")
+    void userIdInTheRequestTest(
+            String caseDescription,
+            String input,
+            boolean valid,
+            String expectedValueOrError
+    ) {
+        checkValidation(
+                input,
+                valid,
+                expectedValueOrError,
+                callbackValidator::hasUserId,
+                expectedValueOrError
+        );
+    }
+
     private static ImmutableMap<String, Object> caseDataWithOcr() {
         return ImmutableMap.of(
                 JOURNEY_CLASSIFICATION, CLASSIFICATION_EXCEPTION,
@@ -310,11 +362,25 @@ class CallbackValidatorTest {
         );
     }
 
-    private <T> void checkValidation(CaseDetails input,
-                                     boolean valid,
-                                     T realValue,
-                                     Function<CaseDetails, Validation<String, ?>> validationMethod,
-                                     String errorString) {
+    private <T> void checkValidation(
+            CaseDetails input,
+            boolean valid,
+            T realValue,
+            Function<CaseDetails, Validation<String, ?>> validationMethod,
+            String errorString
+    ) {
+        Validation<String, ?> validation = validationMethod.apply(input);
+
+        softAssertions(valid, realValue, errorString, validation);
+    }
+
+    private <T> void checkValidation(
+            String input,
+            boolean valid,
+            T realValue,
+            Function<String, Validation<String, ?>> validationMethod,
+            String errorString
+    ) {
         Validation<String, ?> validation = validationMethod.apply(input);
 
         softAssertions(valid, realValue, errorString, validation);
